Reconsider route and lifecycle of each provider
