+++ src/features/session_detail/SessionDetailPage.tsx
@@ -29,7 +29,6 @@ import { WatchedEvents } from "./WatchedEvents"
 import { Description } from "./Description"
 import { Speakers } from "./Speakers"
 import { Note, OndemandVideo } from "@mui/icons-material"
-import { SessionDetailProvider } from "./SessionDetailProvider"
 import { useContext, useEffect } from "react"
 import { SessionDetailContext } from "./SessionDetailContext"
 import { useObservableState } from "observable-hooks"
@@ -39,12 +38,12 @@ import { RequestSubmissionContext } from "../request_submission/RequestSubmissio
 
 export function SessionDetailPage(): JSX.Element {
   return (
-    <SessionDetailProvider>
+    <>
       <Box sx={{ p: 4 }}>
         <SessionDetailBody/>
       </Box>
       <NewRequestFromSessionDialog/>
-    </SessionDetailProvider>      
+    </>
   )
 }
